pyproject.toml
[tool.poetry]
name = "gcp-python-data-samples"
version = "0.1.0"
description = ""
authors = ["Josselin Pujo <pujo.j@sfeir.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
google-cloud-bigquery = "^2.32.0"
gcsfs = "^2022.1.0"
fastavro = "^1.4.9"
dask = "^2022.1.0"
google-cloud-bigquery-storage = {extras = ["fastavro"], version = "^2.11.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
